Move logic to separate hooks, tidy up
