+++ src/shared/ui/share-button/lib/types.ts
@@ -1,0 +1,10 @@
+import React from 'react'
+
+export interface ButtonProps
+  extends React.DetailedHTMLProps<
+      React.ButtonHTMLAttributes<HTMLButtonElement>,
+      HTMLButtonElement
+    >,
+    React.AriaAttributes {
+  isShorten?: boolean
+}
